+++ CHANGELOG.md
@@ -18,6 +18,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.164.0 - 2021-09-02
+
+### Changed
+
+- 口座のtypeOfをAccountで固定
+- PermitのtypeOfをPermitで固定
+
 ## v4.163.1 - 2021-08-31
 
 ### Changed